# openapi-user-auth-service.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Healthcare & RPM User & IAM Authentication Microservice API",
    "description": "### Enterprise-Grade Healthcare & Remote Patient Monitoring (RPM) Identity Platform\n\nThis microservice provides identity lifecycle, Keycloak 24 IAM management, OAuth2 / OIDC Direct Access Grants, HIPAA-compliant RFC 6238 TOTP Multi-Factor Authentication (2FA), Role-Based Access Control (RBAC), composite role permission hierarchies, and physician verification.\n\n#### Key Security & Architecture Features:\n- **Keycloak 24 IAM Integration**: Token issuance, refresh token rotation, single sign-out.\n- **HIPAA Security Rule Enforcement**: Dual-factor TOTP for high-privilege roles, audit logging.\n- **Composite Role Mapping**: Deep hierarchical permission resolution for Doctors, Nurses, Patients, Pharmacists, and Admins.\n- **Group Membership**: Token claims embedding hierarchical paths such as `/Doctors-Writers`.",
    "contact": {
      "name": "Healthcare RPM Engineering & Security Team",
      "url": "https://healthcare.ecosystem.org",
      "email": "linhnv03@ominext.com"
    },
    "license": {
      "name": "HIPAA Audited & Apache 2.0",
      "url": "https://www.hhs.gov/hipaa/for-professionals/security/index.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Direct Local Microservice Runtime (Spring Boot 3.4)"
    },
    {
      "url": "http://localhost:8000/user-service",
...
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication & MFA"],
        "summary": "Register new user account",
        "description": "Provisions a new user account across Keycloak IAM realm and PostgreSQL user_auth_db, assigns realm/client roles, seeds credentials, and sets up optional HIPAA TOTP 2FA.",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered in Keycloak and PostgreSQL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload or validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Username or email already exists in Keycloak IAM",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication & MFA"],
        "summary": "Authenticate user (Direct Access Grant)",
        "description": "Exchanges username/password credentials against Keycloak 24 IAM for a signed RS256 JWT access token and refresh token. Enforces TOTP 2FA verification when configured on user profile.",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful, returns tokens and profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or TOTP verification failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/url": {
      "get": {
        "tags": ["Authentication & MFA"],
        "summary": "Get Keycloak Google Login Authorization URL",
        "description": "Constructs the Keycloak OpenID Connect authorization URL configured with kc_idp_hint=google. Clients (web/mobile/popups) redirect the user to this URL to authenticate with their Google account via Keycloak.",
        "operationId": "getGoogleAuthUrl",
        "parameters": [
          {
            "name": "redirect_uri",
            "in": "query",
            "required": false,
            "description": "Optional client callback redirect URI",
            "schema": {
              "type": "string",
              "example": "http://localhost:3000/auth/callback"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Keycloak Google authorization URL generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleAuthUrlResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/callback": {
      "post": {
        "tags": ["Authentication & MFA"],
        "summary": "Exchange Google OAuth Authorization Code via Keycloak",
        "description": "Exchanges the authorization code returned by Keycloak after Google authentication for signed RS256 JWT access tokens and refresh tokens. Syncs user account and records HIPAA security audit log.",
        "operationId": "exchangeGoogleCode",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Google authentication successful, returns Keycloak JWT tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired authorization code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/token": {
      "post": {
        "tags": ["Authentication & MFA"],
        "summary": "Authenticate with Google ID Token via Keycloak",
        "description": "Authenticates using a Google signed ID token (e.g. from Google Sign-In SDK or One Tap) via Keycloak RFC 8693 Token Exchange or federated verification, automatically provisioning the user in Keycloak and PostgreSQL.",
        "operationId": "loginWithGoogleToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleTokenLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Google Token validated and Keycloak JWT issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid Google Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/config": {
      "get": {
        "tags": ["Authentication & MFA"],
        "summary": "Get Keycloak Google Identity Provider Setup Configuration",
        "description": "Returns the Keycloak broker redirect URI needed for Google Cloud Console OAuth 2.0 Client credentials, the status of Keycloak's Google IDP configuration, and step-by-step setup instructions.",
        "operationId": "getGoogleIdpConfig",
        "parameters": [
          {
            "name": "client_callback_url",
            "in": "query",
            "required": false,
            "description": "Optional client callback URL to test",
            "schema": {
              "type": "string",
              "example": "http://localhost:3000/auth/callback"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Keycloak Google IDP setup instructions and broker endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleIdpConfigResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Authentication & MFA"],
        "summary": "Keycloak Single Sign-Out (Logout)",
        "description": "Revokes the active refresh token and terminates the user's Keycloak session across the realm.",
        "operationId": "logoutUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session terminated and token revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "SUCCESS" },
                    "message": { "type": "string", "example": "Keycloak session terminated and tokens revoked successfully." }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["Authentication & MFA"],
        "summary": "Refresh JWT access token",
        "description": "Uses Keycloak OpenID Connect refresh_token grant to rotate tokens and generate a fresh RS256 access token.",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Expired or invalid refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Authentication & MFA"],
        "summary": "Get current authenticated user profile",
        "description": "Decodes active Keycloak JWT access token and returns subject, claims, expiration, and realm roles.",
        "operationId": "getCurrentUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user token claims and identity metadata"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid Bearer JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/totp/setup": {
      "post": {
        "tags": ["Authentication & MFA"],
        "summary": "Generate TOTP 2FA setup secret and QR code",
        "description": "Generates a cryptographically secure RFC 6238 Base32 secret key and otpauth:// URI for authenticator app enrollment.",
        "operationId": "setupTotp",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "TOTP secret and QR code URI generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "secret": { "type": "string", "example": "JBSWY3DPEHPK3PXP" },
                    "qrCodeUri": { "type": "string", "example": "data:image/png;base64,iVBORw0KGgo..." }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/audit-logs": {
      "get": {
        "tags": ["Authentication & MFA"],
        "summary": "Retrieve HIPAA security audit logs",
        "description": "Queries the immutable PostgreSQL audit log for authentication events, login failures, role updates, and token lifecycle.",
        "operationId": "getAuditLogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of recent HIPAA audit entries"
          },
          "403": {
            "description": "Forbidden - Requires ADMIN realm role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter Keycloak JWT Access Token obtained from `/api/v1/auth/login`. Token carries standard claims (`sub`, `preferred_username`, `realm_access.roles`, `groups`)."
      }
    },
...
  }
}